Ouroboros.Consensus.Mock.Ledger

Documentation

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family Header blk ∷ Type

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

type AddrDist = Map Addr NodeId

Mapping from addresses to node IDs

This is needed in order to assign stake to nodes.

data Addr

Mock address

mkAddrDist ∷ NumCoreNodes → AddrDist

Construct address to node ID mapping

class HasMockTxs a where

Methods

getMockTxs ∷ a → [Tx]

The transactions in the order they are to be applied

data UtxoError

Constructors

MissingInput TxIn

InputOutputMismatch

Fields

Amount

Input

Amount

Output

type Utxo = Map TxIn TxOut

type TxOut = (Addr, Amount)

type TxIn = (TxId, Ix)

type Amount = Word

type Ix = Word

data Tx where

Bundled Patterns

pattern Tx ∷ Expiry → Set TxIn → [TxOut] → Tx

data Expiry

Constructors

DoNotExpire

ExpireAtOnsetOf !SlotNo

txIns ∷ HasMockTxs a ⇒ a → Set TxIn

txOuts ∷ HasMockTxs a ⇒ a → Utxo

confirmed ∷ HasMockTxs a ⇒ a → Set TxId

updateUtxo ∷ HasMockTxs a ⇒ a → Utxo → Except UtxoError Utxo

genesisTx ∷ AddrDist → Tx

Transaction giving initial stake to the nodes

genesisUtxo ∷ AddrDist → Utxo

data MockError blk

Constructors

MockExpired !SlotNo !SlotNo
The transaction expired in the first SlotNo, and it failed to validate in the second SlotNo.

MockUtxoError UtxoError

MockInvalidHash (ChainHash blk) (ChainHash blk)

data MockState blk

Constructors

MockState

Fields

mockUtxo ∷ !Utxo

mockConfirmed ∷ !(Set TxId)

mockTip ∷ !(Point blk)

updateMockState ∷ (GetPrevHash blk, HasMockTxs blk) ⇒ blk → MockState blk → Except (MockError blk) (MockState blk)

updateMockTip ∷ GetPrevHash blk ⇒ Header blk → MockState blk → Except (MockError blk) (MockState blk)

updateMockUTxO ∷ HasMockTxs a ⇒ SlotNo → a → MockState blk → Except (MockError blk) (MockState blk)

genesisMockState ∷ AddrDist → MockState blk

newtype StakeDist

In the mock setup, only core nodes have stake

INVARIANT: The rationals should sum to 1.

Constructors

StakeDist

Fields

stakeDistToMap ∷ Map CoreNodeId Rational

data StakeHolder

Constructors

StakeCore CoreNodeId
Stake of a core node

StakeEverybodyElse
Stake for everybody else (we don't need to distinguish)

stakeWithDefault ∷ Rational → CoreNodeId → StakeDist → Rational

relativeStakes ∷ Map StakeHolder Amount → StakeDist

totalStakes ∷ Map Addr NodeId → Utxo → Map StakeHolder Amount

Compute stakes of all nodes

The Nothing value holds the total stake of all addresses that don't get mapped to a NodeId.

equalStakeDist ∷ AddrDist → StakeDist

Stake distribution where every address has equal state

genesisStakeDist ∷ AddrDist → StakeDist

Genesis stake distribution

data SimpleMockCrypto

data SimpleStandardCrypto

class (HashAlgorithm (SimpleHash c), Typeable c) ⇒ SimpleCrypto c

Associated Types

type SimpleHash c ∷ Type

data SimpleLedgerConfig c ext

Constructors

SimpleLedgerConfig

Fields

simpleMockLedgerConfig ∷ !(MockLedgerConfig c ext)

Config required by the various kinds of mock block (PFT, Praos, ..)

simpleLedgerEraParams ∷ !EraParams

Era parameters

class (SimpleCrypto c, Typeable ext, Show (MockLedgerConfig c ext), NoUnexpectedThunks (MockLedgerConfig c ext)) ⇒ MockProtocolSpecific c ext

Associated Types

type MockLedgerConfig c ext ∷ Type

data SimpleBody

Constructors

SimpleBody

Fields

simpleTxs ∷ [Tx]

data SimpleStdHeader c ext

Constructors

SimpleStdHeader

Fields

simplePrev ∷ ChainHash (SimpleBlock c ext)

simpleSlotNo ∷ SlotNo

simpleBlockNo ∷ BlockNo

simpleBodyHash ∷ Hash (SimpleHash c) SimpleBody

simpleBodySize ∷ Word32

data SimpleBlock' c ext ext'

Constructors

SimpleBlock

Fields

simpleHeader ∷ Header (SimpleBlock' c ext ext')

simpleBody ∷ SimpleBody

type SimpleHeader c ext = Header (SimpleBlock c ext)

type SimpleBlock c ext = SimpleBlock' c ext ext

mkSimpleHeader ∷ SimpleCrypto c ⇒ (ext' → Encoding) → SimpleStdHeader c ext → ext' → Header (SimpleBlock' c ext ext')

matchesSimpleHeader ∷ SimpleCrypto c ⇒ Header (SimpleBlock' c ext ext') → SimpleBlock' c ext ext'' → Bool

Check whether the block matches the header

countSimpleGenTxs ∷ SimpleBlock c ext → Word64

updateSimpleLedgerState ∷ (SimpleCrypto c, Typeable ext) ⇒ SimpleBlock c ext → TickedLedgerState (SimpleBlock c ext) → Except (MockError (SimpleBlock c ext)) (LedgerState (SimpleBlock c ext))

genesisSimpleLedgerState ∷ AddrDist → LedgerState (SimpleBlock c ext)

mkSimpleGenTx ∷ Tx → GenTx (SimpleBlock c ext)

txSize ∷ GenTx (SimpleBlock c ext) → Word32

encodeSimpleHeader ∷ (ext' → Encoding) → Header (SimpleBlock' c ext ext') → Encoding

decodeSimpleHeader ∷ SimpleCrypto c ⇒ (ext' → Encoding) → (∀ s. Decoder s ext') → ∀ s. Decoder s (Header (SimpleBlock' c ext ext'))

simpleBlockBinaryBlockInfo ∷ (SimpleCrypto c, Serialise ext', Typeable ext, Typeable ext') ⇒ SimpleBlock' c ext ext' → BinaryBlockInfo

newtype ForgeExt c ext

Construct the protocol specific part of the block

This is used in forgeSimple, which takes care of the generic part of the mock block.

Note: this is a newtype and not a type class to allow for things in the closure. For example, if Praos had to use a stateful KES key, it could refer to it in its closure.

Constructors

ForgeExt

Fields

forgeExt ∷ TopLevelConfig (SimpleBlock c ext) → IsLeader (BlockProtocol (SimpleBlock c ext)) → SimpleBlock' c ext () → SimpleBlock c ext

forgeSimple

Arguments

∷ (SimpleCrypto c, MockProtocolSpecific c ext)

⇒ ForgeExt c ext

→ TopLevelConfig (SimpleBlock c ext)

→ BlockNo
Current block number

→ SlotNo
Current slot number

→ TickedLedgerState (SimpleBlock c ext)
Current ledger

→ [GenTx (SimpleBlock c ext)]
Txs to include

→ IsLeader (BlockProtocol (SimpleBlock c ext))

→ SimpleBlock c ext

data SignedSimplePBft c c'

Part of the block that gets signed

We just sign the standard header, i.e., without the PBFT extensions. In particular, the signature does not cover the issuer.

The signature does not cover the body explicitly, but since the standard header includes a hash of the body, the signature covers the body implicitly.

Constructors

SignedSimplePBft

Fields

signedSimplePBft ∷ SimpleStdHeader c (SimplePBftExt c c')

newtype SimplePBftExt c c'

Block extension required for PBFT

Constructors

SimplePBftExt

Fields

simplePBftExt ∷ PBftFields c' (SignedSimplePBft c c')

type SimplePBftHeader c c' = SimpleHeader c (SimplePBftExt c c')

Header for PBFT

type SimplePBftBlock c c' = SimpleBlock c (SimplePBftExt c c')

Simple block extended with the fields required for PBFT

c is crypto used for the block itself c' is crypto used for the consensus protocol

forgePBftExt ∷ ∀ c c'. (SimpleCrypto c, PBftCrypto c', Signable (PBftDSIGN c') (SignedSimplePBft c c'), ContextDSIGN (PBftDSIGN c') ~ ()) ⇒ ForgeExt c (SimplePBftExt c c')

data SignedSimpleBft c c'

Part of the block that gets signed

Constructors

SignedSimpleBft

Fields

signedSimpleBft ∷ SimpleStdHeader c (SimpleBftExt c c')

newtype SimpleBftExt c c'

Block extension required for BFT

Constructors

SimpleBftExt

Fields

simpleBftExt ∷ BftFields c' (SignedSimpleBft c c')

type SimpleBftHeader c c' = SimpleHeader c (SimpleBftExt c c')

Header for BFT

type SimpleBftBlock c c' = SimpleBlock c (SimpleBftExt c c')

Simple block extended with the fields required for BFT

c is crypto used for the block itself c' is crypto used for the consensus protocol

forgeBftExt ∷ ∀ c c'. (SimpleCrypto c, BftCrypto c', Signable (BftDSIGN c') (SignedSimpleBft c c')) ⇒ ForgeExt c (SimpleBftExt c c')

data PraosCryptoUnused

newtype SimplePraosRuleExt

Required extension

The WithLeaderSchedule doesn't require anything in the block header. We add the CoreNodeId just so that we can check that the schedule matches the chain.

Constructors

SimplePraosRuleExt

Fields

simplePraosRuleExt ∷ CoreNodeId

type SimplePraosRuleHeader c = SimpleHeader c SimplePraosRuleExt

Header for Proas

type SimplePraosRuleBlock c = SimpleBlock c SimplePraosRuleExt

Simple block extended with the fields required for Praos

c is crypto used for the block itself With an explicit leader schedule we need no crypto for the consensus protocol.

This is an example of a block which is not an instance of SignedBlock.

forgePraosRuleExt ∷ SimpleCrypto c ⇒ ForgeExt c SimplePraosRuleExt

data SignedSimplePraos c c'

Part of the block that gets signed

TODO: Right now we sign all of the extra Praos fields. This may or may not be needed. https://github.com/input-output-hk/cardano-ledger-specs/issues/530 Of course, this Praos is merely a proof of concept so it doesn't really matter either way; we include them here primarily to show that we can.

Constructors

SignedSimplePraos

Fields

signedSimplePraos ∷ SimpleStdHeader c (SimplePraosExt c c')

signedPraosFields ∷ PraosExtraFields c'

newtype SimplePraosExt c c'

Block extension required for Praos

Constructors

SimplePraosExt

Fields

simplePraosExt ∷ PraosFields c' (SignedSimplePraos c c')

type SimplePraosHeader c c' = SimpleHeader c (SimplePraosExt c c')

Header for Proas

type SimplePraosBlock c c' = SimpleBlock c (SimplePraosExt c c')

Simple block extended with the fields required for Praos

c is crypto used for the block itself c' is crypto used for the consensus protocol

forgePraosExt ∷ ∀ c c'. (SimpleCrypto c, PraosCrypto c', Signable (PraosKES c') (SignedSimplePraos c c')) ⇒ HotKey c' → ForgeExt c (SimplePraosExt c c')