Ouroboros.Consensus.Mock.Ledger.Block.BFT

Documentation

data SignedSimpleBft c c'

Part of the block that gets signed

Constructors

SignedSimpleBft

Fields

signedSimpleBft ∷ SimpleStdHeader c (SimpleBftExt c c')

type SimpleBftBlock c c' = SimpleBlock c (SimpleBftExt c c')

Simple block extended with the fields required for BFT

c is crypto used for the block itself c' is crypto used for the consensus protocol

newtype SimpleBftExt c c'

Block extension required for BFT

Constructors

SimpleBftExt

Fields

simpleBftExt ∷ BftFields c' (SignedSimpleBft c c')

type SimpleBftHeader c c' = SimpleHeader c (SimpleBftExt c c')

Header for BFT

forgeBftExt ∷ ∀ c c'. (SimpleCrypto c, BftCrypto c', Signable (BftDSIGN c') (SignedSimpleBft c c')) ⇒ ForgeExt c (SimpleBftExt c c')