Ouroboros.Consensus.Mock.Ledger.Block.PBFT

Documentation

data SignedSimplePBft c c'

Part of the block that gets signed

We just sign the standard header, i.e., without the PBFT extensions. In particular, the signature does not cover the issuer.

The signature does not cover the body explicitly, but since the standard header includes a hash of the body, the signature covers the body implicitly.

Constructors

SignedSimplePBft

Fields

signedSimplePBft ∷ SimpleStdHeader c (SimplePBftExt c c')

type SimplePBftBlock c c' = SimpleBlock c (SimplePBftExt c c')

Simple block extended with the fields required for PBFT

c is crypto used for the block itself c' is crypto used for the consensus protocol

newtype SimplePBftExt c c'

Block extension required for PBFT

Constructors

SimplePBftExt

Fields

simplePBftExt ∷ PBftFields c' (SignedSimplePBft c c')

type SimplePBftHeader c c' = SimpleHeader c (SimplePBftExt c c')

Header for PBFT

forgePBftExt ∷ ∀ c c'. (SimpleCrypto c, PBftCrypto c', Signable (PBftDSIGN c') (SignedSimplePBft c c'), ContextDSIGN (PBftDSIGN c') ~ ()) ⇒ ForgeExt c (SimplePBftExt c c')