Ouroboros.Consensus.Mock.Ledger.Block.Praos

Documentation

data SignedSimplePraos c c'

Part of the block that gets signed

TODO: Right now we sign all of the extra Praos fields. This may or may not be needed. https://github.com/input-output-hk/cardano-ledger-specs/issues/530 Of course, this Praos is merely a proof of concept so it doesn't really matter either way; we include them here primarily to show that we can.

Constructors

SignedSimplePraos

Fields

signedSimplePraos ∷ SimpleStdHeader c (SimplePraosExt c c')

signedPraosFields ∷ PraosExtraFields c'

type SimplePraosBlock c c' = SimpleBlock c (SimplePraosExt c c')

Simple block extended with the fields required for Praos

c is crypto used for the block itself c' is crypto used for the consensus protocol

newtype SimplePraosExt c c'

Block extension required for Praos

Constructors

SimplePraosExt

Fields

simplePraosExt ∷ PraosFields c' (SignedSimplePraos c c')

type SimplePraosHeader c c' = SimpleHeader c (SimplePraosExt c c')

Header for Proas

forgePraosExt ∷ ∀ c c'. (SimpleCrypto c, PraosCrypto c', Signable (PraosKES c') (SignedSimplePraos c c')) ⇒ HotKey c' → ForgeExt c (SimplePraosExt c c')