Ouroboros.Consensus.Mock.Ledger.Stake

Stakeholders

data StakeHolder

Constructors

StakeCore CoreNodeId
Stake of a core node

StakeEverybodyElse
Stake for everybody else (we don't need to distinguish)

Address distribution

type AddrDist = Map Addr NodeId

Mapping from addresses to node IDs

This is needed in order to assign stake to nodes.

Stake distribution

newtype StakeDist

In the mock setup, only core nodes have stake

INVARIANT: The rationals should sum to 1.

Constructors

StakeDist

Fields

stakeDistToMap ∷ Map CoreNodeId Rational

equalStakeDist ∷ AddrDist → StakeDist

Stake distribution where every address has equal state

genesisStakeDist ∷ AddrDist → StakeDist

Genesis stake distribution

relativeStakes ∷ Map StakeHolder Amount → StakeDist

stakeWithDefault ∷ Rational → CoreNodeId → StakeDist → Rational

totalStakes ∷ Map Addr NodeId → Utxo → Map StakeHolder Amount

Compute stakes of all nodes

The Nothing value holds the total stake of all addresses that don't get mapped to a NodeId.

Type family instances

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.