Ouroboros.Consensus.Mock.Ledger.UTxO

Basic definitions

data Addr

Mock address

type Amount = Word

data Expiry

Constructors

DoNotExpire

ExpireAtOnsetOf !SlotNo

type Ix = Word

data Tx where

Bundled Patterns

pattern Tx ∷ Expiry → Set TxIn → [TxOut] → Tx

type TxId = Hash SHA256 Tx

type TxIn = (TxId, Ix)

type TxOut = (Addr, Amount)

type Utxo = Map TxIn TxOut

Computing UTxO

class HasMockTxs a where

Methods

getMockTxs ∷ a → [Tx]

The transactions in the order they are to be applied

data UtxoError

Constructors

MissingInput TxIn

InputOutputMismatch

Fields

Amount

Input

Amount

Output

confirmed ∷ HasMockTxs a ⇒ a → Set TxId

confirmed stands for all the transaction hashes present in the given collection.

txIns ∷ HasMockTxs a ⇒ a → Set TxIn

txOuts ∷ HasMockTxs a ⇒ a → Utxo

updateUtxo ∷ HasMockTxs a ⇒ a → Utxo → Except UtxoError Utxo

Update the Utxo with the transactions from the given a, by removing the inputs and adding the outputs.

Genesis

genesisTx ∷ AddrDist → Tx

Transaction giving initial stake to the nodes

genesisUtxo ∷ AddrDist → Utxo