Ouroboros.Consensus.Mock.Node

Documentation

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

simpleBlockForging ∷ ∀ c ext m. (RunMockBlock c ext, CannotForge (SimpleBlock c ext) ~ Void, ForgeStateInfo (SimpleBlock c ext) ~ (), ForgeStateUpdateError (SimpleBlock c ext) ~ Void, Monad m) ⇒ CanBeLeader (BlockProtocol (SimpleBlock c ext)) → ForgeExt c ext → BlockForging m (SimpleBlock c ext)

Can be used when CanBeLeader is static

Orphan instances

(LedgerSupportsProtocol (SimpleBlock SimpleMockCrypto ext), BlockSupportsProtocol (SimpleBlock SimpleMockCrypto ext), Show (CannotForge (SimpleBlock SimpleMockCrypto ext)), Show (ForgeStateInfo (SimpleBlock SimpleMockCrypto ext)), Show (ForgeStateUpdateError (SimpleBlock SimpleMockCrypto ext)), Serialise ext, RunMockBlock SimpleMockCrypto ext) ⇒ RunNode (SimpleBlock SimpleMockCrypto ext)

Instance details

Methods

nodeBlockFetchSize ∷ Header (SimpleBlock SimpleMockCrypto ext) → SizeInBytes

nodeImmutableDbChunkInfo ∷ TopLevelConfig (SimpleBlock SimpleMockCrypto ext) → ChunkInfo

nodeCheckIntegrity ∷ TopLevelConfig (SimpleBlock SimpleMockCrypto ext) → SimpleBlock SimpleMockCrypto ext → Bool

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig (SimpleBlock SimpleMockCrypto ext) → InitChainDB m (SimpleBlock SimpleMockCrypto ext) → m ()

HasNetworkProtocolVersion (SimpleBlock SimpleMockCrypto ext)

Instance details

Associated Types

type BlockNodeToNodeVersion (SimpleBlock SimpleMockCrypto ext) ∷ Type

type BlockNodeToClientVersion (SimpleBlock SimpleMockCrypto ext) ∷ Type

SupportedNetworkProtocolVersion (SimpleBlock SimpleMockCrypto ext)

Instance details

Methods

supportedNodeToNodeVersions ∷ Proxy (SimpleBlock SimpleMockCrypto ext) → Map NodeToNodeVersion (BlockNodeToNodeVersion (SimpleBlock SimpleMockCrypto ext))

supportedNodeToClientVersions ∷ Proxy (SimpleBlock SimpleMockCrypto ext) → Map NodeToClientVersion (BlockNodeToClientVersion (SimpleBlock SimpleMockCrypto ext))