Ouroboros.Consensus.Mock.Node.Abstract

Documentation

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

class (MockProtocolSpecific c ext, EncodeDisk (SimpleBlock c ext) (ChainDepState (BlockProtocol (SimpleBlock c ext))), DecodeDisk (SimpleBlock c ext) (ChainDepState (BlockProtocol (SimpleBlock c ext)))) ⇒ RunMockBlock c ext where

Protocol specific functionality required to run consensus with mock blocks

Methods

mockNetworkMagic ∷ BlockConfig (SimpleBlock c ext) → NetworkMagic

constructMockNetworkMagic ∷ HasCallStack ⇒ NetworkMagic

Construct protocol magic ID depending on where in the code this is called

The sole purpose of this is to make sure that these mock protocols have different IDs from each other and from regular protocols.

Orphan instances

RunMockBlock c ext ⇒ ConfigSupportsNode (SimpleBlock c ext)

Instance details

Methods

getSystemStart ∷ BlockConfig (SimpleBlock c ext) → SystemStart

getNetworkMagic ∷ BlockConfig (SimpleBlock c ext) → NetworkMagic