Ouroboros.Consensus.Mock.Protocol.Praos

Description

Proof of concept implementation of Praos

Documentation

data Praos c

data PraosFields c toSign

The fields that Praos required in the header

Constructors

PraosFields

Fields

praosSignature ∷ SignedKES (PraosKES c) toSign

praosExtraFields ∷ PraosExtraFields c

data PraosExtraFields c

Fields that should be included in the signature

Constructors

PraosExtraFields

Fields

praosCreator ∷ CoreNodeId

praosRho ∷ CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType)

praosY ∷ CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType)

data PraosParams

Praos parameters that are node independent

Constructors

PraosParams

Fields

praosLeaderF ∷ !Double

praosSecurityParam ∷ !SecurityParam

praosSlotsPerEpoch ∷ !Word64

praosLifetimeKES ∷ !Natural

newtype PraosChainDepState c

Constructors

PraosChainDepState

Fields

praosHistory ∷ [BlockInfo c]

data HotKey c

Constructors

HotKey

Fields

!Period

Absolute period of the KES key

!(SignKeyKES (PraosKES c))

HotKeyPoisoned

forgePraosFields ∷ (PraosCrypto c, Signable (PraosKES c) toSign, HasCallStack) ⇒ PraosProof c → HotKey c → (PraosExtraFields c → toSign) → PraosFields c toSign

newtype HotKeyEvolutionError

The HotKey could not be evolved to the given Period.

Constructors

HotKeyEvolutionError Period

evolveKey ∷ PraosCrypto c ⇒ SlotNo → HotKey c → (HotKey c, UpdateInfo (HotKey c) (HotKey c) HotKeyEvolutionError)

To be used in conjunction with, e.g., updateMVar.

NOTE: when the key's period is after the target period, we shouldn't use it, but we currently do. In real TPraos we check this in tpraosCheckCanForge.

Tags

class (KESAlgorithm (PraosKES c), VRFAlgorithm (PraosVRF c), HashAlgorithm (PraosHash c), Typeable c, Typeable (PraosVRF c), Condense (SigKES (PraosKES c)), Signable (PraosVRF c) (Natural, SlotNo, VRFType), ContextKES (PraosKES c) ~ (), ContextVRF (PraosVRF c) ~ ()) ⇒ PraosCrypto (c ∷ Type)

Associated Types

type PraosKES c ∷ Type

type PraosVRF c ∷ Type

type PraosHash c ∷ Type

data PraosStandardCrypto

data PraosMockCrypto

data PraosValidateView c

Constructors

Signable (PraosKES c) signed ⇒ PraosValidateView (PraosFields c signed) signed

praosValidateView ∷ (SignedHeader hdr, Signable (PraosKES c) (Signed hdr)) ⇒ (hdr → PraosFields c (Signed hdr)) → hdr → PraosValidateView c

Convenience constructor for PraosValidateView

data PraosValidationError c

Constructors

PraosInvalidSlot SlotNo SlotNo

PraosUnknownCoreId CoreNodeId

PraosInvalidSig String (VerKeyKES (PraosKES c)) Natural (SigKES (PraosKES c))

PraosInvalidCert (VerKeyVRF (PraosVRF c)) (Natural, SlotNo, VRFType) Natural (CertVRF (PraosVRF c))

PraosInsufficientStake Double Natural

Type instances

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data BlockInfo c

Constructors

BlockInfo

Fields

biSlot ∷ !SlotNo

biRho ∷ !(CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType))

biStake ∷ !StakeDist

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.