Ouroboros.Consensus.Mock.Ledger.Block

Description

Simple block to go with the mock ledger

None of the definitions in this module depend on, or even refer to, any specific consensus protocols.

Documentation

type SimpleBlock c ext = SimpleBlock' c ext ext

type SimpleHeader c ext = Header (SimpleBlock c ext)

data SimpleBlock' c ext ext'

Constructors

SimpleBlock

Fields

simpleHeader ∷ Header (SimpleBlock' c ext ext')

simpleBody ∷ SimpleBody

data family Header blk ∷ Type

data SimpleStdHeader c ext

Constructors

SimpleStdHeader

Fields

simplePrev ∷ ChainHash (SimpleBlock c ext)

simpleSlotNo ∷ SlotNo

simpleBlockNo ∷ BlockNo

simpleBodyHash ∷ Hash (SimpleHash c) SimpleBody

simpleBodySize ∷ Word32

data SimpleBody

Constructors

SimpleBody

Fields

simpleTxs ∷ [Tx]

type family SimpleHash c ∷ Type

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

Working with SimpleBlock

mkSimpleHeader ∷ SimpleCrypto c ⇒ (ext' → Encoding) → SimpleStdHeader c ext → ext' → Header (SimpleBlock' c ext ext')

matchesSimpleHeader ∷ SimpleCrypto c ⇒ Header (SimpleBlock' c ext ext') → SimpleBlock' c ext ext'' → Bool

Check whether the block matches the header

countSimpleGenTxs ∷ SimpleBlock c ext → Word64

Configuration

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data SimpleLedgerConfig c ext

Constructors

SimpleLedgerConfig

Fields

simpleMockLedgerConfig ∷ !(MockLedgerConfig c ext)

Config required by the various kinds of mock block (PFT, Praos, ..)

simpleLedgerEraParams ∷ !EraParams

Era parameters

Protocol-specific part

class (SimpleCrypto c, Typeable ext, Show (MockLedgerConfig c ext), NoUnexpectedThunks (MockLedgerConfig c ext)) ⇒ MockProtocolSpecific c ext

Associated Types

type MockLedgerConfig c ext ∷ Type

UpdateLedger

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

updateSimpleLedgerState ∷ (SimpleCrypto c, Typeable ext) ⇒ SimpleBlock c ext → TickedLedgerState (SimpleBlock c ext) → Except (MockError (SimpleBlock c ext)) (LedgerState (SimpleBlock c ext))

genesisSimpleLedgerState ∷ AddrDist → LedgerState (SimpleBlock c ext)

ApplyTx (mempool support)

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

mkSimpleGenTx ∷ Tx → GenTx (SimpleBlock c ext)

txSize ∷ GenTx (SimpleBlock c ext) → Word32

Crypto

class (HashAlgorithm (SimpleHash c), Typeable c) ⇒ SimpleCrypto c

data SimpleStandardCrypto

data SimpleMockCrypto

Serialisation

encodeSimpleHeader ∷ (ext' → Encoding) → Header (SimpleBlock' c ext ext') → Encoding

decodeSimpleHeader ∷ SimpleCrypto c ⇒ (ext' → Encoding) → (∀ s. Decoder s ext') → ∀ s. Decoder s (Header (SimpleBlock' c ext ext'))

simpleBlockBinaryBlockInfo ∷ (SimpleCrypto c, Serialise ext', Typeable ext, Typeable ext') ⇒ SimpleBlock' c ext ext' → BinaryBlockInfo