Ouroboros.Consensus.Mock.Ledger.Block

Description

Simple block to go with the mock ledger

None of the definitions in this module depend on, or even refer to, any specific consensus protocols.

Documentation

data family BlockQuery blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

data family Header blk

type SimpleBlock c ext = SimpleBlock' c ext ext

data SimpleBlock' c ext ext'

Constructors

SimpleBlock

Fields

simpleHeader ∷ Header (SimpleBlock' c ext ext')

simpleBody ∷ SimpleBody

data SimpleBody

Constructors

SimpleBody

Fields

simpleTxs ∷ [Tx]

type family SimpleHash c ∷ Type

type SimpleHeader c ext = Header (SimpleBlock c ext)

data SimpleStdHeader c ext

Constructors

SimpleStdHeader

Fields

simplePrev ∷ ChainHash (SimpleBlock c ext)

simpleSlotNo ∷ SlotNo

simpleBlockNo ∷ BlockNo

simpleBodyHash ∷ Hash (SimpleHash c) SimpleBody

simpleBodySize ∷ Word32

Working with SimpleBlock

countSimpleGenTxs ∷ SimpleBlock c ext → Word64

matchesSimpleHeader ∷ SimpleCrypto c ⇒ Header (SimpleBlock' c ext ext') → SimpleBlock' c ext ext'' → Bool

Check whether the block matches the header

mkSimpleHeader ∷ SimpleCrypto c ⇒ (ext' → Encoding) → SimpleStdHeader c ext → ext' → Header (SimpleBlock' c ext ext')

Create a header by hashing the header without hash and adding to the resulting value.

Configuration

data family BlockConfig blk

Static configuration required to work with this type of blocks

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data SimpleLedgerConfig c ext

Constructors

SimpleLedgerConfig

Fields

simpleMockLedgerConfig ∷ !(MockLedgerConfig c ext)

Config required by the various kinds of mock block (PFT, Praos, ..)

simpleLedgerEraParams ∷ !EraParams

Era parameters

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

Protocol-specific part

class (SimpleCrypto c, Typeable ext, Show (MockLedgerConfig c ext), NoThunks (MockLedgerConfig c ext)) ⇒ MockProtocolSpecific c ext

Associated Types

type MockLedgerConfig c ext ∷ Type

UpdateLedger

data family LedgerState blk

Ledger state associated with a block

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

genesisSimpleLedgerState ∷ AddrDist → LedgerState (SimpleBlock c ext)

updateSimpleLedgerState ∷ (SimpleCrypto c, Typeable ext) ⇒ SimpleBlock c ext → TickedLedgerState (SimpleBlock c ext) → Except (MockError (SimpleBlock c ext)) (LedgerState (SimpleBlock c ext))

ApplyTx (mempool support)

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx

A generalized transaction, GenTx, identifier.

data family Validated x

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

mkSimpleGenTx ∷ Tx → GenTx (SimpleBlock c ext)

txSize ∷ GenTx (SimpleBlock c ext) → Word32

Crypto

class (HashAlgorithm (SimpleHash c), Typeable c) ⇒ SimpleCrypto c

data SimpleMockCrypto

data SimpleStandardCrypto

Serialisation

decodeSimpleHeader ∷ SimpleCrypto c ⇒ (ext' → Encoding) → (∀ s. Decoder s ext') → ∀ s. Decoder s (Header (SimpleBlock' c ext ext'))

encodeSimpleHeader ∷ (ext' → Encoding) → Header (SimpleBlock' c ext ext') → Encoding

simpleBlockBinaryBlockInfo ∷ (SimpleCrypto c, Serialise ext', Typeable ext, Typeable ext') ⇒ SimpleBlock' c ext ext' → BinaryBlockInfo