Ouroboros.Consensus.Mock.Ledger.Block.BFT

Documentation

type SimpleBftBlock c c' = SimpleBlock c (SimpleBftExt c c')

Simple block extended with the fields required for BFT

c is crypto used for the block itself c' is crypto used for the consensus protocol

type SimpleBftHeader c c' = SimpleHeader c (SimpleBftExt c c')

Header for BFT

newtype SimpleBftExt c c'

Block extension required for BFT

Constructors

SimpleBftExt

Fields

simpleBftExt ∷ BftFields c' (SignedSimpleBft c c')

data SignedSimpleBft c c'

Part of the block that gets signed

Constructors

SignedSimpleBft

Fields

signedSimpleBft ∷ SimpleStdHeader c (SimpleBftExt c c')