Ouroboros.Consensus.Mock.Ledger.Block.PraosRule

Description

Test the Praos chain selection rule (with explicit leader schedule)

Documentation

type SimplePraosRuleBlock c = SimpleBlock c SimplePraosRuleExt

Simple block extended with the fields required for Praos

c is crypto used for the block itself With an explicit leader schedule we need no crypto for the consensus protocol.

This is an example of a block which is not an instance of SignedBlock.

newtype SimplePraosRuleExt

Required extension

The WithLeaderSchedule doesn't require anything in the block header. We add the CoreNodeId just so that we can check that the schedule matches the chain.

Constructors

SimplePraosRuleExt

Fields

simplePraosRuleExt ∷ CoreNodeId

type SimplePraosRuleHeader c = SimpleHeader c SimplePraosRuleExt

Header for Proas

data PraosCryptoUnused