Ouroboros.Consensus.Mock.Ledger.Stake

Stakeholders

data StakeHolder

Constructors

StakeCore CoreNodeId
Stake of a core node

StakeEverybodyElse
Stake for everybody else (we don't need to distinguish)

Address distribution

type AddrDist = Map Addr NodeId

Mapping from addresses to node IDs

This is needed in order to assign stake to nodes.

Stake distribution

newtype StakeDist

In the mock setup, only core nodes have stake

Constructors

StakeDist

Fields

stakeDistToMap ∷ Map CoreNodeId Rational

stakeWithDefault ∷ Rational → CoreNodeId → StakeDist → Rational

relativeStakes ∷ Map StakeHolder Amount → StakeDist

totalStakes ∷ Map Addr NodeId → Utxo → Map StakeHolder Amount

Compute stakes of all nodes

The Nothing value holds the total stake of all addresses that don't get mapped to a NodeId.

equalStakeDist ∷ AddrDist → StakeDist

Stake distribution where every address has equal state

genesisStakeDist ∷ AddrDist → StakeDist

Genesis stake distribution