Ouroboros.Consensus.Mock.Node

Documentation

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

simpleBlockForging ∷ ∀ c ext m. (RunMockBlock c ext, CannotForge (SimpleBlock c ext) ~ Void, ForgeStateInfo (SimpleBlock c ext) ~ (), ForgeStateUpdateError (SimpleBlock c ext) ~ Void, Monad m) ⇒ CanBeLeader (BlockProtocol (SimpleBlock c ext)) → ForgeExt c ext → BlockForging m (SimpleBlock c ext)

Can be used when CanBeLeader is static

Orphan instances

(LedgerSupportsProtocol (SimpleBlock SimpleMockCrypto ext), Show (CannotForge (SimpleBlock SimpleMockCrypto ext)), Show (ForgeStateInfo (SimpleBlock SimpleMockCrypto ext)), Show (ForgeStateUpdateError (SimpleBlock SimpleMockCrypto ext)), Serialise ext, RunMockBlock SimpleMockCrypto ext) ⇒ RunNode (SimpleBlock SimpleMockCrypto ext)

Instance details

NodeInitStorage (SimpleBlock SimpleMockCrypto ext)

Instance details

Methods

nodeImmutableDbChunkInfo ∷ StorageConfig (SimpleBlock SimpleMockCrypto ext) → ChunkInfo

nodeCheckIntegrity ∷ StorageConfig (SimpleBlock SimpleMockCrypto ext) → SimpleBlock SimpleMockCrypto ext → Bool

nodeInitChainDB ∷ IOLike m ⇒ StorageConfig (SimpleBlock SimpleMockCrypto ext) → InitChainDB m (SimpleBlock SimpleMockCrypto ext) → m ()

BlockSupportsMetrics (SimpleBlock c ext)

Instance details

Methods

isSelfIssued ∷ BlockConfig (SimpleBlock c ext) → Header (SimpleBlock c ext) → WhetherSelfIssued

HasNetworkProtocolVersion (SimpleBlock SimpleMockCrypto ext)

Instance details

Associated Types

type BlockNodeToNodeVersion (SimpleBlock SimpleMockCrypto ext)

type BlockNodeToClientVersion (SimpleBlock SimpleMockCrypto ext)

SupportedNetworkProtocolVersion (SimpleBlock SimpleMockCrypto ext)

Instance details

Methods

supportedNodeToNodeVersions ∷ Proxy (SimpleBlock SimpleMockCrypto ext) → Map NodeToNodeVersion (BlockNodeToNodeVersion (SimpleBlock SimpleMockCrypto ext))

supportedNodeToClientVersions ∷ Proxy (SimpleBlock SimpleMockCrypto ext) → Map NodeToClientVersion (BlockNodeToClientVersion (SimpleBlock SimpleMockCrypto ext))

latestReleasedNodeVersion ∷ Proxy (SimpleBlock SimpleMockCrypto ext) → (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)