Ouroboros.Consensus.Mock.Node.Abstract

Documentation

class MockProtocolSpecific c ext ⇒ RunMockBlock c ext where

Protocol specific functionality required to run consensus with mock blocks

Methods

forgeExt ∷ MonadRandom m ⇒ TopLevelConfig (SimpleBlock c ext) → Update m (NodeState (SimpleBlock c ext)) → IsLeader (BlockProtocol (SimpleBlock c ext)) → SimpleBlock' c ext () → m (SimpleBlock c ext)

Construct the protocol specific part of the block

This is used in forgeSimple, which takes care of the generic part of the mock block.

mockProtocolMagicId ∷ TopLevelConfig (SimpleBlock c ext) → ProtocolMagicId

mockEncodeConsensusState ∷ TopLevelConfig (SimpleBlock c ext) → ConsensusState (BlockProtocol (SimpleBlock c ext)) → Encoding

mockDecodeConsensusState ∷ TopLevelConfig (SimpleBlock c ext) → Decoder s (ConsensusState (BlockProtocol (SimpleBlock c ext)))

constructMockProtocolMagicId ∷ HasCallStack ⇒ ProtocolMagicId

Construct protocol magic ID depending on where in the code this is called

The sole purpose of this is to make sure that these mock protocols have different IDs from each other and from regular protocols.