Ouroboros.Consensus.Mock.Protocol.Praos

Description

Proof of concept implementation of Praos

Documentation

data HotKey c

The key used for the given period or a stub Poisoned value.

A key will be poisoned if it failed to evolve by updateKES, and will remain poisoned forever after that.

Constructors

HotKey

Fields

!Period

Absolute period of the KES key

!(SignKeyKES (PraosKES c))

HotKeyPoisoned

newtype HotKeyEvolutionError

The HotKey could not be evolved to the given Period.

Constructors

HotKeyEvolutionError Period

data Praos c

An uninhabited type representing the Praos protocol.

newtype PraosChainDepState c

The chain dependent state, in this case as it is a mock, we just will store a list of BlockInfos that allow us to look into the past.

Constructors

PraosChainDepState

Fields

praosHistory ∷ [BlockInfo c]

newtype PraosEvolvingStake

An association from epoch to stake distributions.

Should be used when checking if someone is the leader of a particular slot. This is sufficiently good for a mock protocol as far as consensus is concerned. It is not strictly necessary that the stake distribution is computed from previous epochs, as we just need to consider that:

1) an attacker cannot influence it. 2) all the nodes agree on the same value for each Slot.

Each pair stores the stake distribution established by the end of the epoch in the first item of the pair. See latestEvolvedStakeDistAsOfEpoch for the intended interface.

If no value is returned, that means we are checking the stake before any changes have happened so we should consult instead the praosInitialStake.

Constructors

PraosEvolvingStake (Map EpochNo StakeDist)

data PraosExtraFields c

Fields that should be included in the signature

Constructors

PraosExtraFields

Fields

praosCreator ∷ CoreNodeId

praosRho ∷ CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType)

praosY ∷ CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType)

data PraosFields crypto typeBeingSigned

The fields that Praos required in the header

Constructors

PraosFields

Fields

praosSignature ∷ SignedKES (PraosKES crypto) typeBeingSigned

praosExtraFields ∷ PraosExtraFields crypto

data PraosParams

Praos parameters that are node independent

Constructors

PraosParams

Fields

praosLeaderF ∷ !Double

f, the active slots coefficient, defined in 3.3 in the Praos paper.

praosSecurityParam ∷ !SecurityParam

k, maximum number of blocks we can rollback

praosSlotsPerEpoch ∷ !Word64

R, slots in each epoch, defined in section 3 in the Praos paper.

emptyPraosEvolvingStake ∷ PraosEvolvingStake

evolveKey ∷ PraosCrypto c ⇒ SlotNo → HotKey c → (HotKey c, UpdateInfo (HotKey c) HotKeyEvolutionError)

To be used in conjunction with, e.g., updateMVar.

NOTE: when the key's period is after the target period, we shouldn't use it, but we currently do. In real TPraos we check this in tpraosCheckCanForge.

forgePraosFields ∷ (PraosCrypto c, Signable (PraosKES c) toSign, HasCallStack) ⇒ PraosProof c → HotKey c → (PraosExtraFields c → toSign) → PraosFields c toSign

Create a PraosFields using a proof, a key and the data to be signed.

It is done by signing whatever is extracted from the extra fields by mkToSign and storing the signature and the extra fields on a PraosFields.

Tags

class (KESAlgorithm (PraosKES c), VRFAlgorithm (PraosVRF c), HashAlgorithm (PraosHash c), Typeable c, Typeable (PraosVRF c), Condense (SigKES (PraosKES c)), Signable (PraosVRF c) (Natural, SlotNo, VRFType), ContextKES (PraosKES c) ~ (), ContextVRF (PraosVRF c) ~ ()) ⇒ PraosCrypto (c ∷ Type)

Associated Types

type PraosKES c ∷ Type

type PraosVRF c ∷ Type

type PraosHash c ∷ Type

data PraosMockCrypto

data PraosStandardCrypto

data PraosValidateView c

A validate view is an association from the (signed) value to the PraosFields that contains the signature that sign it.

In this mock implementation, this could have been simplified to use SignedSimplePraos but from the consensus point of view, it is not relevant which actual value is being signed, that's why we use the existential.

Constructors

∀ signed.Signable (PraosKES c) signed ⇒ PraosValidateView (PraosFields c signed) signed

data PraosValidationError c

An error that can arise during validation

Constructors

PraosInvalidSlot SlotNo SlotNo

PraosUnknownCoreId CoreNodeId

PraosInvalidSig String (VerKeyKES (PraosKES c)) Natural (SigKES (PraosKES c))

PraosInvalidCert (VerKeyVRF (PraosVRF c)) (Natural, SlotNo, VRFType) Natural (CertVRF (PraosVRF c))

PraosInsufficientStake Double Natural

praosValidateView ∷ (SignedHeader hdr, Signable (PraosKES c) (Signed hdr)) ⇒ (hdr → PraosFields c (Signed hdr)) → hdr → PraosValidateView c

Convenience constructor for PraosValidateView

Type instances

data BlockInfo c

Constructors

BlockInfo

Fields

biSlot ∷ !SlotNo

biRho ∷ !(CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType))

data family ConsensusConfig p

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.