Ouroboros.Consensus.Mock.Protocol.Praos

Description

Proof of concept implementation of Praos

Documentation

data Praos c

data PraosFields c toSign

The fields that Praos required in the header

Constructors

PraosFields

Fields

praosSignature ∷ SignedKES (PraosKES c) toSign

praosExtraFields ∷ PraosExtraFields c

data PraosExtraFields c

Fields that should be included in the signature

Constructors

PraosExtraFields

Fields

praosCreator ∷ CoreNodeId

praosRho ∷ CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType)

praosY ∷ CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType)

data PraosParams

Praos parameters that are node independent

Constructors

PraosParams

Fields

praosLeaderF ∷ !Double

praosSecurityParam ∷ !SecurityParam

praosSlotsPerEpoch ∷ !Word64

praosLifetimeKES ∷ !Natural

data PraosNodeState c

Constructors

PraosKeyAvailable !(SignKeyKES (PraosKES c))
The KES key is available

PraosKeyEvolving
The KES key is being evolved by another thread

Any thread that sees this value should back off and retry.

forgePraosFields ∷ (MonadRandom m, PraosCrypto c, Signable (PraosKES c) toSign) ⇒ ConsensusConfig (Praos c) → Update m (PraosNodeState c) → PraosProof c → (PraosExtraFields c → toSign) → m (PraosFields c toSign)

Tags

class (KESAlgorithm (PraosKES c), VRFAlgorithm (PraosVRF c), HashAlgorithm (PraosHash c), Typeable c, Typeable (PraosVRF c), Condense (SigKES (PraosKES c)), Signable (PraosVRF c) (Natural, SlotNo, VRFType), ContextKES (PraosKES c) ~ (), ContextVRF (PraosVRF c) ~ ()) ⇒ PraosCrypto (c ∷ ★)

Associated Types

type PraosKES c ∷ ★

type PraosVRF c ∷ ★

type PraosHash c ∷ ★

data PraosStandardCrypto

data PraosMockCrypto

data PraosValidateView c

Constructors

Signable (PraosKES c) signed ⇒ PraosValidateView SlotNo (PraosFields c signed) signed

praosValidateView ∷ (HasHeader hdr, SignedHeader hdr, Signable (PraosKES c) (Signed hdr)) ⇒ (hdr → PraosFields c (Signed hdr)) → hdr → PraosValidateView c

Convenience constructor for PraosValidateView

data PraosValidationError c

Constructors

PraosInvalidSlot SlotNo SlotNo

PraosUnknownCoreId CoreNodeId

PraosInvalidSig String (VerKeyKES (PraosKES c)) Natural (SigKES (PraosKES c))

PraosInvalidCert (VerKeyVRF (PraosVRF c)) (Natural, SlotNo, VRFType) Natural (CertVRF (PraosVRF c))

PraosInsufficientStake Double Natural

Type instances

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data BlockInfo c

Constructors

BlockInfo

Fields

biSlot ∷ !SlotNo

biRho ∷ !(CertifiedVRF (PraosVRF c) (Natural, SlotNo, VRFType))

biStake ∷ !StakeDist